views/club: Merge into "clubs"
